Title: 10개 입력받기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231202/10개 입력받기/receive-10-inputs.c
@@ -2,7 +2,7 @@
 
 int main() {
     int arr[10], cnt = 0, sum = 0;
-    double avg;
+    float avg;
     for(int i = 0;i < 10;i++) {
         scanf("%d", &arr[i]);
         if(arr[i] == 0) {
@@ -12,6 +12,6 @@ int main() {
         sum += arr[i];
     }
     avg = sum / cnt;
-    printf("%d %.1lf", sum, avg);
+    printf("%d %.1f", sum, avg);
     return 0;
 }
